Fix skill() crashing with large scales.
In the worst case (Spellcasting at 26.99 with a -5 apt), it'd overflow with
a scale as low as 92.

This commit doesn't actually allows unlimited values of scale, merely gives
us another 16 bits.  Max scale of 5976516 should be enough for everybody,
right?
(cherry picked from commit 0dbff5aea1c62227886adc593fa2838ccfecf047)

+++ crawl-ref/source/skills2.cc
@@ -146,7 +146,9 @@ int get_skill_progress(skill_type sk, int level, int points, int scale)
 
     const int needed = skill_exp_needed(level + 1, sk);
     const int prev_needed = skill_exp_needed(level, sk);
-    const int amt_done = points - prev_needed;
+    // A scale as small as 92 would overflow with 31 bits if skill_rdiv()
+    // is involved: needed can be 91985, skill_rdiv() multiplies by 256.
+    const int64_t amt_done = points - prev_needed;
     int prog = (amt_done * scale) / (needed - prev_needed);
 
     ASSERT(prog >= 0);
